// src/lib/geometry/geometry.enum.ts
export enum DirectionEnum {
	CW = 'cw',
	CCW = 'ccw'
}

export enum MirrorEnum {
	HORIZONTAL = 'horizontal',
	VERTICAL = 'vertical'
}

export enum GeometryTypeEnum {
	SEGMENT = 'segment',
	ARC = 'arc',
	CIRCLE = 'circle',
	POINT = 'point',
	POLYSHAPE = 'polyshape',
	QUADRATIC_CURVE = 'quadratic_curve',
	CUBIC_CURVE = 'cubic_curve',
	ELLIPSE = 'ellipse',
	RECTANGLE = 'rectangle'
}

export enum OriginEnum {
	MIDDLE_CENTER = 'middle-center',
	BOTTOM_LEFT = 'bottom-left',
	BOTTOM_RIGHT = 'bottom-right',
	TOP_LEFT = 'top-left',
	TOP_RIGHT = 'top-right'
}